+++ Runtime/SettingsAsset/SettingsAsset.cs
@@ -6,16 +6,16 @@ namespace Kryz.Settings
 {
 	public abstract class SettingsAsset : ScriptableObject
 	{
-		[SerializeField, ReadOnly] uint id;
+		[SerializeField, ReadOnly] ulong id;
 		[SerializeField, ReadOnly] string settingName; // Used to serialize a human readable name.
 		[SerializeField] bool enabled = true;
 
-		public uint Id => id;
+		public ulong Id => id;
 		public string Name => settingName;
 		public bool Enabled => enabled;
 
 		[Conditional("UNITY_EDITOR")]
-		internal void SetIdAndName(uint id, string name)
+		internal void SetIdAndName(ulong id, string name)
 		{
 			this.id = id;
 			settingName = name;
